feat: add a minDistance configuration option
This change makes it possible to configure a minimum distance that is used as a threshold to determine whether to begin a pan or zoom operation when triggered via pointer events.

This option can help in situations when small movements during a click/tap cause the element to move slightly which is a little off putting.